+++ example/src/scenarios/get-attestation.ts
@@ -4,7 +4,7 @@ import {
   createCryptoContextFor,
   type IntegrityContext,
 } from "@pagopa/io-react-native-wallet";
-import { error, result } from "../types";
+import { error, result } from "./types";
 import { WALLET_PROVIDER_BASE_URL } from "@env";
 
 /**
